Patch 0.1.3: Refactor blog page into page -> grid -> component
